backend setup for using social logins
