+++ src/server/auth/auth.router.js
@@ -1,21 +1,31 @@
 import { Router } from 'express';
-import passport from 'passport';
-import { handleLogin, authFacebook, authFacebookCallback, authGoogle,
-authGoogleCallback, authTwitter, authTwitterCallback, authGithub, authGithubCallback } from '../api/auth/authenticator';
-import { isAuthenticated } from './auth.service';
+import {
+  authFacebook,
+  authFacebookCallback,
+  authGoogle,
+  authGoogleCallback,
+  authTwitter,
+  authTwitterCallback,
+  authGithub,
+  authGithubCallback
+} from './providers';
+import { ensureAuthenticated, unlink } from './authenticator';
 import * as ctrl from './auth.controller';
-
 
 const router = Router();
 
-router.post('/login', handleLogin);
-router.post('/signup', ctrl.signUp);
+router.post('/login', ctrl.handleLogin);
+router.post('/signup', ctrl.handleSignup);
 router.post('/logout', ctrl.logout);
 router.post('/forgot', ctrl.forgottenPassword);
 router.post('/reset/:token', ctrl.resetPassword);
 
-router.route('/check')
-  .get(isAuthenticated(), ctrl.checkUser);
+router.get('/logout', ctrl.logout);
+
+router.route('/check').get(ensureAuthenticated(), ctrl.checkUser);
+
+router.delete('/account', ensureAuthenticated, ctrl.accountDelete);
+router.get('/unlink/:provider', ensureAuthenticated, unlink);
 
 router.post('/google', authGoogle);
 router.get('/google/callback', authGoogleCallback);
@@ -26,14 +36,5 @@ router.get('/twitter/callback', authTwitterCallback);
 router.post('/github', authGithub);
 router.get('/github/callback', authGithubCallback);
 
-// router.get('/facebook', passport.authenticate('facebook', {
-//   scope: ['email']
-// }));
-//
-// router.get('/facebook/callback',
-//   passport.authenticate('facebook', {
-//     successRedirect: '/',
-//     failureRedirect: '/login'
-//   })
-// );
+
 export default router;
